Add MIS, normal mapping, and Russian roulette
Sampled distributions used in MIS are currently only the specular and diffuse lobes. Changed how textures are kept in memory and how they are processed (WIP). GLTF roughness/metalness maps now work properly. Cleaned up a bit in the shaders (WIP). Added environment light/map and light parsing to scene description (currently only the environment light colour does anything in the render). The ray origins are now offset in the direction of the true normal of the hit geometry.